Factor out the rest of the tuners
